feat: linked project drag and drop to get all projects for next semester
