+++ src/mocks/Projects.mock.tsx
@@ -8,8 +8,8 @@ export const mockProjects: ProjectDetailsType[] = [
     projectTitle: 'Chronobiology',
     projectClientDetails: {
       id: '',
-      firstName: 'John', 
-      lastName: 'Doe', 
+      firstName: 'John',
+      lastName: 'Doe',
       role: UserRole.Client,
       updatedAt: '',
       createdAt: '',
@@ -23,28 +23,30 @@ export const mockProjects: ProjectDetailsType[] = [
       'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
     availableResources:
       'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-    specialRequirements: '',
-    numberOfTeams: "Up to 2 teams",
+    specialRequirements: 'No',
+    numberOfTeams: 'Up to 2 teams',
     futureConsideration: true,
-    semesters: [{
-      id: "semester-2-2025",
-      deadline: "02-05-2024",
-      name: "Semester 2 2025",
-      startDate: '02-05-2024',
-      endDate: '02-05-2024',
-      updatedAt: '02-05-2024',
-      createdAt: '02-05-2024'
-    }],
+    semesters: [
+      {
+        id: 'semester-2-2025',
+        deadline: '02-05-2024',
+        name: 'Semester 2 2025',
+        startDate: '02-05-2024',
+        endDate: '02-05-2024',
+        updatedAt: '02-05-2024',
+        createdAt: '02-05-2024',
+      },
+    ],
     submittedDate: new Date('2025-06-30'),
   },
   {
-    semesterProjectId: "sp3",
+    semesterProjectId: 'sp3',
     projectId: 'p2',
     projectTitle: 'Capstone Submission',
     projectClientDetails: {
       id: '',
-      firstName: 'Jane', 
-      lastName: 'Doe', 
+      firstName: 'Jane',
+      lastName: 'Doe',
       role: UserRole.Client,
       updatedAt: '',
       createdAt: '',
@@ -58,18 +60,20 @@ export const mockProjects: ProjectDetailsType[] = [
       'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
     availableResources:
       'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.',
-    specialRequirements: '',
+    specialRequirements: 'No',
     numberOfTeams: 'Up to two teams',
     futureConsideration: true,
-    semesters: [{
-      id: "semester-2-2025",
-      deadline: "02-05-2024",
-      name: "Semester 2 2025",
-      startDate: '02-05-2024',
-      endDate: '02-05-2024',
-      updatedAt: '02-05-2024',
-      createdAt: '02-05-2024'
-    }],
+    semesters: [
+      {
+        id: 'semester-2-2025',
+        deadline: '02-05-2024',
+        name: 'Semester 2 2025',
+        startDate: '02-05-2024',
+        endDate: '02-05-2024',
+        updatedAt: '02-05-2024',
+        createdAt: '02-05-2024',
+      },
+    ],
     submittedDate: new Date('2025-03-20'),
   },
 ]
